feat!: extract Schematic schema layer into lightweight gotcha_core crate
BREAKING CHANGE: workspace bumped to 0.3.0.

Introduce a new lightweight `gotcha_core` crate holding the `Schematic`
trait, `EnhancedSchema`, and all data-type impls, so a crate can derive and
use `Schematic` without pulling in axum/tokio/tower. `gotcha` re-exports
everything (`use gotcha::Schematic;` etc. keep working). The `Schematic`
derive now expands to `::gotcha_core::` paths.

`Responsible` and `ParameterProvider` also move into `gotcha_core`, behind
an optional, off-by-default `axum` feature that `gotcha` enables. They each
carry a blanket impl over `Schematic` that overlaps specific impls
(`Json<T>`, `Path<(T1,)>`); the compiler can only prove non-overlap in the
crate where `Schematic` is defined, so they must be co-located with it.
`gotcha_core`'s default dependency tree stays free of axum/regex/either.

Migration: any crate that derives `Schematic` must add `gotcha_core` to its
own Cargo.toml (extern prelude only sees direct deps) — one line, no code
change.
