+++ examples/extension/Cargo.toml
@@ -5,6 +5,8 @@ edition = "2021"
 
 [dependencies]
 gotcha = { path = "../../gotcha", features = ["openapi"] }
+# Required because `#[derive(Schematic)]` expands to `::gotcha_core::` paths.
+gotcha_core = { path = "../../gotcha_core" }
 tokio = { version = "1", features = ["full"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
