+++ examples/openapi/Cargo.toml
@@ -6,7 +6,9 @@ version = "0.1.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-gotcha = {version="0.2", path = "../../gotcha", features = ["openapi", "http1"]}
+gotcha = {version="0.3", path = "../../gotcha", features = ["openapi", "http1"]}
+# Required because `#[derive(Schematic)]` expands to `::gotcha_core::` paths.
+gotcha_core = { path = "../../gotcha_core" }
 tokio = {version = "1", features = ["macros", 'rt-multi-thread']}
 serde = {version="1", features=["derive"]}
 serde_json = "1"
