+++ examples/task/Cargo.toml
@@ -6,6 +6,6 @@ version = "0.1.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-gotcha = {version="0.2", path = "../../gotcha", features = ["task", "http1"]}
+gotcha = {version="0.3", path = "../../gotcha", features = ["task", "http1"]}
 tokio = {version = "1", features = ["macros", 'rt-multi-thread']}
 serde = {version="1", features=["derive"]}
