+++ gotcha_macro/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "gotcha_macro"
 description = "gotcha web framework macro"
-version = "0.2.0"
+version = "0.3.0"
 edition = "2021"
 license = "MIT"
